Added in report format and download button to Clim Explorer and PA Explorer. Colors for XY Plot are properly fixed this time in both the PA and Clim explorers.
Still need to sort out the best way to do the PAs. Right now, the subselections based on view mess with the ability to compare polygons. This should be simply solvable by just plotting a subsection, but creating the selected polygons by subsetting the complete list of polygons, rather than the view subset.